prepare admin system for control over site products and content
ensure unsaved cloud images are deleted after Product.save()

use simple-icons lib for updated icons for social links

add slide counter for carousel on product item page

+++ routes/product.js
@@ -1,5 +1,6 @@
 const router = require('express').Router();
 const Product = require('../models/Product');
+const isAuthed = require('../modules/auth-check');
 
 router.get('/:product_name', async (req, res, next) => {
     const param = req.params.product_name.trim().replace(/\-|\$/g, "\\W+");
@@ -9,9 +10,9 @@ router.get('/:product_name', async (req, res, next) => {
     res.render("product-item", { product });
 });
 
-router.post('/create', async (req, res) => {
+router.post('/create', isAuthed, async (req, res) => {
     const { name, info, colour, length_inches, size, price, unit_stock_qty, image_file, image_url } = req.body;
-    const images = [image_file, image_url].flat().filter(e => e);
+    const images = [image_file, image_url].flat().filter(e => e).map(url => ({ url }));
     const units = [{ colour, length_inches, size, price, unit_stock_qty, images }];
     try {
         await Product.create({ name, info, units });
@@ -21,32 +22,45 @@ router.post('/create', async (req, res) => {
     }
 });
 
-router.post('/unit/create', async (req, res) => {
-    const { product_id, colour, length_inches, size, price, unit_stock_qty, image_file, image_url } = req.body;
-    const images = [image_file, image_url].flat().filter(e => e);
+router.post('/update', isAuthed, async (req, res) => {
+    const { product_id, unit_id, name, info, colour, length_inches, size, price, unit_stock_qty, image_file, image_url } = req.body;
+    const images = [image_file, image_url].flat().filter(e => e).map(url => ({ url }));
     try {
         const product = await Product.findById(product_id);
-        product.units.shift({ colour, length_inches, size, price, unit_stock_qty, images });
+        const populated_fields = [colour, length_inches, size, price, unit_stock_qty].filter(e => e);
+        const new_unit = unit_id == "new";
+
+        if (name) product.name = name;
+        if (info) product.info = info;
+
+        if (new_unit && populated_fields.length) {
+            product.units.shift({ colour, length_inches, size, price, unit_stock_qty, images });
+        }
+        else if (unit_id) {
+            const unit = product.units.find(u => unit_id == u.id);
+            if (unit) {
+                if (colour) unit.colour = colour;
+                if (length_inches) unit.length_inches = length_inches;
+                if (size) unit.size = size;
+                if (price) unit.price = price;
+                if (unit_stock_qty) unit.unit_stock_qty = unit_stock_qty;
+                unit.images = images;
+            }
+        }
+
         await product.save();
-        res.send("Product unit saved successfully!");
+        res.send(`Product saved successfully${new_unit ? ", with the new unit" : ""}!`);
     } catch (err) {
         res.status(400).send(err.message);
     }
 });
 
-router.post('/delete', async (req, res) => {
+router.post('/delete', isAuthed, async (req, res) => {
+    if (!req.body.product_id && !req.body.unit_id) return res.status(400).send("Nothing selected");
     try {
         await Product.deleteByIds(req.body.product_id);
-        res.send("Product now removed from stock");
-    } catch (err) {
-        res.status(400).send(err.message);
-    }
-});
-
-router.post('/unit/delete', async (req, res) => {
-    try {
         await Product.deleteUnitsByIds(req.body.unit_id);
-        res.send("Product unit now removed from stock");
+        res.send("Selected products now removed from stock");
     } catch (err) {
         res.status(400).send(err.message);
     }
@@ -58,7 +72,7 @@ router.post('/get-price', async (req, res) => {
 
     const product = await Product.findById(id).catch(e => null);
     const unit = product?.units.find(u => u.unit_description === style);
-    if (!unit) return res.status(404).send("Sorry, your preferred style is not available right now.\nPlease check again later.");
+    if (!unit) return res.status(404).send("Sorry, your selected style option is not available right now.\nPlease check again later.");
 
     res.send({ price: unit.price, stock_qty: unit.unit_stock_qty });
 });
